requirements.txt
ulauncher.api